add button styling, and delete button renders list

+++ models/teamModel.js
@@ -8,7 +8,7 @@ module.exports = function(sequelize, DataTypes) {
       allowNull: false,
     },
     key: {
-      type: DataTypes.INTEGER(20),
+      type: DataTypes.BIGINT(20),
       allowNull: false,
       unique: true
     }
@@ -30,14 +30,5 @@ module.exports = function(sequelize, DataTypes) {
   }
 
 
-  // Creating a custom method for our User model. This will check if an unhashed password entered by the user can be compared to the hashed password stored in our database
-  // Team.prototype.validPassword = function(password) {
-  //   return bcrypt.compareSync(password, this.password);
-  // };
-  // Hooks are automatic methods that run during various phases of the User Model lifecycle
-  // In this case, before a User is created, we will automatically hash their password
-  // Team.addHook("beforeCreate", function(team) {
-  //   team.password = bcrypt.hashSync(team.password, bcrypt.genSaltSync(10), null);
-  // });
   return Team;
 };
